// src/app/store/actionTypes/users-action.types.ts
export enum UsersActiontypes {
    GET_USERS = '[Users] get users',
    GET_USERS_OK = '[Users] get users ok',
    GET_USERS_FAIL = '[Users] get users fail',
    CLEAR_GET_USERS_ERROR = '[Users] clear get users error',
    CLEAR_ALL_USERS = '[Users] clear all',

    GET_SEARCHED_USERS = '[Users] get searched users',
    GET_SEARCHED_USERS_OK = '[Users] get searched users ok',
    GET_SEARCHED_USERS_FAIL = '[Users] get searched users fail',
    CLEAR_GET_SEARCHED_USERS_ERROR = '[Users] clear get searched users error',
    CLEAR_SEARCHED_USERS = '[Users] clear all searched users'
}